updated rake sample_data

+++ lib/tasks/dev.rake
@@ -6,12 +6,25 @@ task({ :sample_data => :environment }) do
     ActiveRecord::Base.connection.reset_pk_sequence!(t)
   end
 
-  Board.destroy_all
-  Post.destroy_all
+  if Rails.env.development?
+    User.destroy_all
+    Board.destroy_all
+    Post.destroy_all
+  end
+
+  usernames = ["alice", "bob", "carol", "dave", "eve"]
+
+  usernames.each do |username|
+    user = User.new
+    user.email = "#{username}@example.com"
+    user.password = "password"
+    user.save  
+  end
   
   5.times do
     board = Board.new
     board.name = Faker::Address.community
+    board.user_id = User.all.sample.id
     board.save
 
     rand(10..50).times do
@@ -21,10 +34,12 @@ task({ :sample_data => :environment }) do
       post.body = Faker::Lorem.paragraphs(number: rand(1..5), supplemental: true).join("\n\n")
       post.created_at = Faker::Date.backward(days: 120)
       post.expires_on = post.created_at + rand(3..90).days
+      post.user_id = User.all.sample.id
       post.save
     end
   end
 
+  puts "There are now #{User.count} rows in the users table."
   puts "There are now #{Board.count} rows in the boards table."
   puts "There are now #{Post.count} rows in the posts table."
 end
